Rename some .h files to .hpp
