+++ src/guiengine/CGUISpriteBank.hpp
