+++ src/guiengine/widgets/CGUIEditBox.cpp
@@ -2,7 +2,7 @@
 // This file is part of the "Irrlicht Engine".
 // For conditions of distribution and use, see copyright notice in irrlicht.h
 
-#include "CGUIEditBox.h"
+#include "guiengine/widgets/CGUIEditBox.hpp"
 
 #include "IGUISkin.h"
 #include "IGUIEnvironment.h"
