+++ src/guiengine/widgets/CGUISTKListBox.cpp
@@ -3,7 +3,7 @@
 // This file is part of the "Irrlicht Engine".
 // For conditions of distribution and use, see copyright notice in irrlicht.h
 
-#include "guiengine/widgets/CGUISTKListBox.h"
+#include "guiengine/widgets/CGUISTKListBox.hpp"
 
 #include "IGUISkin.h"
 #include "IGUIEnvironment.h"
