+++ src/guiengine/widgets/list_widget.cpp
@@ -17,14 +17,14 @@
 
 #include "guiengine/widgets/list_widget.hpp"
 
-#include "guiengine/CGUISpriteBank.h"
+#include "guiengine/CGUISpriteBank.hpp"
 #include "guiengine/engine.hpp"
 #include "io/file_manager.hpp"
 
 #include <IGUIElement.h>
 #include <IGUISkin.h>
 #include <IGUIEnvironment.h>
-#include "IGUIFontBitmap.h"
+#include <IGUIFontBitmap.h>
 
 #include <sstream>
