+++ src/states_screens/options_screen_input2.cpp
@@ -18,7 +18,7 @@
 #include "states_screens/options_screen_input2.hpp"
 
 #include "config/user_config.hpp"
-#include "guiengine/CGUISpriteBank.h"
+#include "guiengine/CGUISpriteBank.hpp"
 #include "guiengine/scalable_font.hpp"
 #include "guiengine/screen.hpp"
 #include "guiengine/widget.hpp"
